accessible/tests/mochitest/tree/test_select.html

<!DOCTYPE html>
<html>
<head>
  <title>HTML select control tests</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      var accTree = {
        role: ROLE_LISTBOX,
        children: [
          {
            role: ROLE_GROUPING,
            children: [
              {
                role: ROLE_STATICTEXT,
                children: [ ]
              },
              {
                role: ROLE_OPTION,
                children: [
                  {
                    role: ROLE_TEXT_LEAF
                  }
                ]
              },
              {
                role: ROLE_OPTION,
                children: [
                  {
                    role: ROLE_TEXT_LEAF
                  }
                ]
              }
            ]
          },
          {
            role: ROLE_OPTION,
            children: [
              {
                role: ROLE_TEXT_LEAF
              }
            ]
          }
        ]
      };
      testAccessibleTree("listbox", accTree);

      accTree = {
        role: ROLE_COMBOBOX,
        children: [
          {
            role: ROLE_COMBOBOX_LIST,
            children: [
              {
                role: ROLE_GROUPING,
                children: [
                  {
                    role: ROLE_STATICTEXT,
                    children: [ ]
                  },
                  {
                    role: ROLE_COMBOBOX_OPTION,
                    children: [
                      {
                        role: ROLE_TEXT_LEAF
                      }
                    ]
                  },
                  {
                    role: ROLE_COMBOBOX_OPTION,
                    children: [
                      {
                        role: ROLE_TEXT_LEAF
                      }
                    ]
                  },
                ]
              },
              {
                role: ROLE_COMBOBOX_OPTION,
                children: [
                  {
                    role: ROLE_TEXT_LEAF
                  }
                ]
              }
            ]
          }
        ]
      };
      testAccessibleTree("combobox", accTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="remove all the code in #ifdef COMBO_BOX_WITH_THREE_CHILDREN"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=506616">
    Mozilla Bug 506616
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <select id="listbox" size="4">
    <optgroup label="Colors">
      <option>Red</option>
      <option>Blue</option>
    </optgroup>
    <option>Animal</option>
  </select>

  <select id="combobox">
    <optgroup label="Colors">
      <option>Red</option>
      <option>Blue</option>
    </optgroup>
    <option>Animal</option>
  </select>
</body>
</html>
